Update to version 1.2.0

+++ source/container/load-test.sh
@@ -6,6 +6,7 @@ UUID=$(cat /proc/sys/kernel/random/uuid)
 echo "S3_BUCKET:: ${S3_BUCKET}"
 echo "TEST_ID:: ${TEST_ID}"
 echo "TEST_TYPE:: ${TEST_TYPE}"
+echo "FILE_TYPE:: ${FILE_TYPE}"
 echo "PREFIX:: ${PREFIX}"
 echo "UUID ${UUID}"
 
@@ -14,16 +15,39 @@ aws s3 cp s3://$S3_BUCKET/test-scenarios/$TEST_ID.json test.json
 
 # download JMeter jmx file
 if [ "$TEST_TYPE" != "simple" ]; then
-  aws s3 cp s3://$S3_BUCKET/public/test-scenarios/$TEST_TYPE/$TEST_ID.jmx ./
+  # Copy *.jar to JMeter library path. See the Taurus JMeter path: https://gettaurus.org/docs/JMeter/
+  JMETER_LIB_PATH=`find ~/.bzt/jmeter-taurus -type d -name "lib"`
+  echo "cp $PWD/*.jar $JMETER_LIB_PATH"
+  cp $PWD/*.jar $JMETER_LIB_PATH
+
+  if [ "$FILE_TYPE" != "zip" ]; then
+    aws s3 cp s3://$S3_BUCKET/public/test-scenarios/$TEST_TYPE/$TEST_ID.jmx ./
+  else
+    aws s3 cp s3://$S3_BUCKET/public/test-scenarios/$TEST_TYPE/$TEST_ID.zip ./
+    unzip $TEST_ID.zip
+    # only looks for the first jmx file.
+    JMETER_SCRIPT=`find . -name "*.jmx" | head -n 1`
+    if [ -z "$JMETER_SCRIPT" ]; then
+      echo "There is no JMeter script in the zip file."
+      exit 1
+    fi
+
+    sed -i -e "s|$TEST_ID.jmx|$JMETER_SCRIPT|g" test.json
+  fi
 fi
 
 echo "Running test"
-bzt test.json -o modules.console.disable=true
+bzt test.json -o modules.console.disable=true | tee -a result.tmp
+CALCULATED_DURATION=`cat result.tmp | grep -m1 "Test duration" | awk -F ' ' '{ print $5 }' | awk -F ':' '{ print ($1 * 3600) + ($2 * 60) + $3 }'`
 
 # upload custom results to S3 if any
 # every file goes under $TEST_ID/$PREFIX/$UUID to distinguish the result correctly
 if [ "$TEST_TYPE" != "simple" ]; then
-  cat $TEST_ID.jmx | grep filename > results.txt
+  if [ "$FILE_TYPE" != "zip" ]; then
+    cat $TEST_ID.jmx | grep filename > results.txt
+  else
+    cat $JMETER_SCRIPT | grep filename > results.txt
+  fi
   sed -i -e 's/<stringProp name="filename">//g' results.txt
   sed -i -e 's/<\/stringProp>//g' results.txt
   sed -i -e 's/ //g' results.txt
@@ -43,5 +67,17 @@ if [ "$TEST_TYPE" != "simple" ]; then
   done
 fi
 
-echo "Uploading results"
-aws s3 cp /tmp/artifacts/results.xml s3://$S3_BUCKET/results/${TEST_ID}/${PREFIX}-${UUID}.xml
+if [ -f /tmp/artifacts/results.xml ]; then
+  echo "Validating Test Duration"
+  TEST_DURATION=`xmlstarlet sel -t -v "/FinalStatus/TestDuration" /tmp/artifacts/results.xml`
+
+  if (( $(echo "$TEST_DURATION > $CALCULATED_DURATION" | bc -l) )); then
+    echo "Updating test duration: $CALCULATED_DURATION s"
+    xmlstarlet ed -L -u /FinalStatus/TestDuration -v $CALCULATED_DURATION /tmp/artifacts/results.xml
+  fi
+
+  echo "Uploading results"
+  aws s3 cp /tmp/artifacts/results.xml s3://$S3_BUCKET/results/${TEST_ID}/${PREFIX}-${UUID}.xml
+else
+  echo "There might be an error happened while the test."
+fi
